+++ src/utils.ts
@@ -4,6 +4,7 @@ import * as File from './utils/file.js';
 import * as Etc from './utils/etc.js';
 import * as Timeout from './utils/timeout.js';
 import * as Fetch from './utils/fetch.js';
+import * as Notify from './utils/notify.js';
 
 export const USER = GLib.get_user_name();
 export const CACHE_DIR = `${GLib.get_user_cache_dir()}/${pkg.name.split('.').pop()}`;
@@ -29,10 +30,13 @@ export const lookUpIcon = Etc.lookUpIcon;
 
 export const fetch = Fetch.fetch;
 
+export const notify = Notify.notify;
+
 export default {
     exec, execAsync, subprocess,
     readFile, readFileAsync, writeFile, monitorFile,
     timeout, interval, idle,
     loadInterfaceXML, bulkConnect, bulkDisconnect, ensureDirectory, lookUpIcon,
     fetch,
+    notify,
 };